Improve RedSound stream play matching

+++ src/RedSound/RedStream.cpp
@@ -479,7 +479,7 @@ int StreamPlay(int streamID, void* streamHeader, int fileSize, int pan, int volu
 		adpcmSampleOffset = REDSOUND_STREAM_FILE_AUDIO_OFFSET;
 		streamFile = reinterpret_cast<RedStreamFile*>(streamHeader);
 		adpcmHeader = RedStreamFileGetAdpcm(streamFile, REDSOUND_STREAM_LEFT_CHANNEL);
-		adpcmHeader->m_data.pred_scale = (short)RedStreamFileGetSampleByte(streamFile, adpcmSampleOffset);
+		adpcmHeader->m_data.pred_scale = RedStreamFileGetSampleByte(streamFile, adpcmSampleOffset);
 		adpcmHeader->m_data.yn1 = adpcmHeader->m_data.yn2 = 0;
 		if (streamData->m_header.m_channelCount == REDSOUND_STREAM_STEREO_CHANNEL_COUNT) {
 			if (streamData->m_header.m_loopStart < REDSOUND_STREAM_LOOP_ENABLED_MIN) {
@@ -488,7 +488,7 @@ int StreamPlay(int streamID, void* streamHeader, int fileSize, int pan, int volu
 				adpcmSampleOffset += REDSOUND_STREAM_STEREO_RIGHT_FRAME_OFFSET;
 			}
 			RedStreamAdpcmHeaderGetChannel(adpcmHeader, REDSOUND_STREAM_RIGHT_CHANNEL)->m_data.pred_scale =
-			    (short)RedStreamFileGetSampleByte(streamFile, adpcmSampleOffset);
+			    RedStreamFileGetSampleByte(streamFile, adpcmSampleOffset);
 			RedStreamAdpcmHeaderGetChannel(adpcmHeader, REDSOUND_STREAM_RIGHT_CHANNEL)->m_data.yn1 =
 			    RedStreamAdpcmHeaderGetChannel(adpcmHeader, REDSOUND_STREAM_RIGHT_CHANNEL)->m_data.yn2 = 0;
 		}
